etl/bronze/requirements.txt
cassandra-driver
cassandra-sigv4

